+++ mod-source-record-manager-server/src/main/java/org/folio/services/util/EventHandlingUtil.java
@@ -12,12 +12,15 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.kafka.KafkaConfig;
 import org.folio.kafka.KafkaTopicNameHelper;
+import org.folio.kafka.SimpleKafkaProducerManager;
+import org.folio.kafka.services.KafkaProducerRecordBuilder;
 import org.folio.processing.events.utils.PomReaderUtil;
 import org.folio.rest.jaxrs.model.Event;
 import org.folio.rest.jaxrs.model.EventMetadata;
 import org.folio.rest.jaxrs.model.RecordCollection;
 import org.folio.rest.tools.utils.ModuleName;
 import org.folio.services.exceptions.RecordsPublishingException;
+
 import static org.folio.services.util.RecordConversionUtil.RECORDS;
 
 import java.util.List;
@@ -56,19 +59,20 @@ public final class EventHandlingUtil {
     String producerName = eventType + "_Producer";
     LOGGER.debug("sendEventToKafka:: Starting to send event to Kafka for eventType: {} and recordId: {} and chunkId: {}", eventType, recordId, chunkId);
 
-    KafkaProducer<String, String> producer =
-      KafkaProducer.createShared(Vertx.currentContext().owner(), producerName, kafkaConfig.getProducerProps());
-    producer.write(record, war -> {
-      producer.end(ear -> producer.close());
-      if (war.succeeded()) {
+    KafkaProducer<String, String> producer = createProducer(eventType, kafkaConfig);
+    producer.send(record)
+      .<Void>mapEmpty()
+      .eventually(x -> producer.close())
+      .onSuccess(res -> {
         logSendingSucceeded(eventType, chunkId, recordId);
         promise.complete(true);
-      } else {
-        Throwable cause = war.cause();
+      })
+      .onFailure(err -> {
+        Throwable cause = err.getCause();
         LOGGER.warn("{} write error for event {}:", producerName, eventType, cause);
         handleKafkaPublishingErrors(promise, eventPayload, cause);
-      }
-    });
+      });
+
     return promise.future();
   }
 
@@ -89,7 +93,12 @@ public final class EventHandlingUtil {
   }
 
   public static KafkaProducerRecord<String, String> createProducerRecord(Event event, String key, String topicName, List<KafkaHeader> kafkaHeaders) {
-    KafkaProducerRecord<String, String> record = KafkaProducerRecord.create(topicName, key, Json.encode(event));
+    KafkaProducerRecord<String, String> record = new KafkaProducerRecordBuilder<String, Object>(event.getEventMetadata().getTenantId())
+      .key(key)
+      .value(event)
+      .topic(topicName)
+      .build();
+
     record.addHeaders(kafkaHeaders);
     return record;
   }
@@ -116,8 +125,7 @@ public final class EventHandlingUtil {
   }
 
   public static KafkaProducer<String, String> createProducer(String eventType, KafkaConfig kafkaConfig) {
-    String producerName = eventType + "_Producer";
-    return KafkaProducer.createShared(Vertx.currentContext().owner(), producerName, kafkaConfig.getProducerProps());
+    return new SimpleKafkaProducerManager(Vertx.currentContext().owner(), kafkaConfig).createShared(eventType);
   }
 
   private static void handleKafkaPublishingErrors(Promise<Boolean> promise, String eventPayload, Throwable cause) {